+++ products/DesignerApp.browser.js
@@ -63,8 +63,8 @@ class DesignerCommander extends AbstractCommander {
   }
   _toIceTray(program) {
     const columns = program.getProgramWidth()
-    const cellTypes = new jtree.TreeNode(program.getInPlaceCellTypeTreeWithNodeConstructorNames())
-    const rootCellTypes = new jtree.TreeNode(program.getInPlacePreludeCellTypeTreeWithNodeConstructorNames())
+    const cellTypes = new jtree.TreeNode(program.toCellTypeTreeWithNodeConstructorNames())
+    const rootCellTypes = new jtree.TreeNode(program.toPreludeCellTypeTreeWithNodeConstructorNames())
     const table = program
       .getProgramAsCells()
       .map((line, lineIndex) => {
@@ -472,7 +472,6 @@ class tableComponent extends AbstractTreeComponent {
     type checkbox
     value compileCommand
     class onCodeUp
-    checked
    a Compile
     ${WillowConstants.DataShadowEvents.onClickCommand} compileCommand
    span  |
@@ -480,7 +479,6 @@ class tableComponent extends AbstractTreeComponent {
     type checkbox
     value visualizeCommand
     class onCodeUp
-    checked
    a Explain
     ${WillowConstants.DataShadowEvents.onClickCommand} visualizeCommand
    textarea