+++ products/TreeComponentFramework.browser.js
@@ -980,7 +980,7 @@ class AbstractTreeComponent extends jtree.GrammarBackedNode {
     if (commandMethod.includes(" "))
       // todo: cleanup
       commandMethod = commandMethod.split(" ")[0]
-    this.addToCommandLog([commandMethod, params.uno, params.dos].filter(item => item).join(" "))
+    this.addToCommandLog([commandMethod, params.uno, params.dos].filter(identity => identity).join(" "))
     this._onCommandWillRun() // todo: remove. currently used by ohayo
     let treeComponent = stumpNode.getStumpNodeTreeComponent()
     let commander = treeComponent.getCommander()