+++ products/commandLineApp.node.js
@@ -34,7 +34,7 @@ class CommandLineApp {
       dir = Utils.getParentFolder(dir)
     }
     if (!Disk.exists(filePath)) throw new Error(`No '${filePath}' found.`)
-    return execSync([filePath, buildCommandName, argument].filter(commandWord => commandWord).join(" "), { encoding: "utf8" })
+    return execSync([filePath, buildCommandName, argument].filter(identity => identity).join(" "), { encoding: "utf8" })
   }
   combine(grammarName) {
     const content = this.programs(grammarName)
@@ -186,6 +186,16 @@ ${errors.length} errors found ${errors.length ? "\n" + errors.join("\n") : ""}`
   }
   allHistory() {
     return this._getHistoryFile()
+  }
+  webForm(grammarName, nodeTypeId) {
+    // webForm grammarName nodeTypeId? Build a web form for the given grammar
+    const grammarPath = this._getGrammarPathByGrammarNameOrThrow(grammarName)
+    const grammarProgram = new jtree.GrammarProgram(Disk.read(grammarPath)).getRootConstructor()
+    let def = new grammarProgram().getDefinition()
+    if (nodeTypeId) def = def.getNodeTypeDefinitionByNodeTypeId("chargeNode")
+    const stumpCode = def.toStumpString()
+    const stumpNode = require("../products/stump.nodejs.js")
+    return new stumpNode(stumpCode).compile()
   }
   _getHistoryFile() {
     Disk.createFileIfDoesNotExist(this._getLogFilePath(), "command paramOne paramTwo timestamp\n")