+++ treeBase/TreeBase.ts
@@ -41,7 +41,7 @@ class TreeBaseFile extends TreeNode {
         const nodes = this.findNodes(this._getFilePath() + " " + term)
         return nodes.map((node: treeNotationTypes.treeNode) => node.childrenToString()).join("\n")
       })
-      .filter(a => a)
+      .filter(identity => identity)
       .join("\n")
   }